# src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: E-Commerce API
  version: 1.0.0
  description: API for a simple e-commerce platform

servers:
  - url: http://localhost:8080

paths:
  /api/products:
    get:
      summary: Retrieve all products
      operationId: getAllProducts
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              example: |
                [
                  {
                    "id": 1,
                    "name": "Product A",
                    "price": 20.99
                  },
                  {
                    "id": 2,
                    "name": "Product B",
                    "price": 30.49
                  }
                ]
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'

    post:
      summary: Create a new product
      operationId: addProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              example: |
                {
                  "id": 3,
                  "name": "New Product",
                  "price": 25.99
                }
              schema:
                $ref: '#/components/schemas/ProductResponse'

  /api/products/{id}:
    get:
      summary: Retrieve a product by ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              example: |
                {
                  "id": 1,
                  "name": "Product A",
                  "price": 20.99
                }
              schema:
                $ref: '#/components/schemas/ProductResponse'

    put:
      summary: Update an existing product
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              example: |
                {
                  "id": 1,
                  "name": "Updated Product",
                  "price": 22.99
                }
              schema:
                $ref: '#/components/schemas/ProductResponse'

    delete:
      summary: Delete a product by ID
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Product deleted successfully

components:
  schemas:
    ProductRequest:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
      required:
        - name
        - price

    ProductResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
      required:
        - id
        - name
        - price